move airquality component into overview component
